Refactored code to no longer cache pia conv_in, hopefully fixing all existing PIA issues
